Consolidates tests
